deduplicate `tool` spec
