+++ cyclonedx-bom/src/specs/common/mod.rs
@@ -26,3 +26,4 @@ pub(crate) mod organization;
 pub(crate) mod property;
 pub(crate) mod service;
 pub(crate) mod signature;
+pub(crate) mod tool;
