+++ cyclonedx-bom/src/specs/v1_3/mod.rs
@@ -21,6 +21,6 @@ pub(crate) mod component;
 pub(crate) mod composition;
 pub(crate) mod license;
 pub(crate) mod metadata;
-pub(crate) mod tool;
 
 pub(crate) use crate::specs::common::service::v1_3 as service;
+pub(crate) use crate::specs::common::tool::v1_3 as tool;
