+++ cyclonedx-bom/src/specs/v1_4/mod.rs
@@ -21,7 +21,6 @@ pub(crate) mod component;
 pub(crate) mod composition;
 pub(crate) mod license;
 pub(crate) mod metadata;
-pub(crate) mod tool;
 pub(crate) mod vulnerability;
 pub(crate) mod vulnerability_analysis;
 pub(crate) mod vulnerability_credits;
@@ -31,3 +30,4 @@ pub(crate) mod vulnerability_source;
 pub(crate) mod vulnerability_target;
 
 pub(crate) use crate::specs::common::service::v1_4 as service;
+pub(crate) use crate::specs::common::tool::v1_4 as tool;
