+++ cyclonedx-bom/src/specs/v1_5/mod.rs
@@ -23,7 +23,6 @@ pub(crate) mod composition;
 pub(crate) mod license;
 pub(crate) mod lifecycles;
 pub(crate) mod metadata;
-pub(crate) mod tool;
 pub(crate) mod vulnerability;
 pub(crate) mod vulnerability_analysis;
 pub(crate) mod vulnerability_credits;
@@ -34,3 +33,4 @@ pub(crate) mod vulnerability_target;
 
 #[allow(unused_imports)]
 pub(crate) use crate::specs::common::service::v1_5 as service;
+pub(crate) use crate::specs::common::tool::v1_5 as tool;
